net/ngbe: disable LLDP by default
In the new firmware versions, LLDP is enabled by default to implement
new features in other drivers. But it is useless in DPDK. So disable
it in device initialization to prevent it from affecting hardware
default behavior.

Build errors on some ARM platforms are introduced by this patch:

../drivers/net/ngbe/base/ngbe_mng.c: In function 'ngbe_hic_get_lldp':
../drivers/net/ngbe/base/ngbe_mng.c:127:36: error: array subscript 2
is outside array bounds of 'struct ngbe_hic_write_lldp[1]' [-Werror=
array-bounds]
127 |                         buffer[bi] = rd32a(hw, NGBE_MNGMBX, bi);
../drivers/net/ngbe/base/ngbe_mng.c:385:36: note: while referencing
'buffer'
385 |         struct ngbe_hic_write_lldp buffer;

So remove the redundant code 'resp->cmd == 0x30' to fix this issue.

Signed-off-by: Jiawen Wu <jiawenwu@trustnetic.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>