common/mlx5: remove unneeded field when modify RQ table
The attribute "rqt_max_size" of the RQT(RQ table) context is only
needed by when creating a new one. When modifying an old one, there
is no need to set this attribute.

The firmware usually ignored this unneeded field. When enabling some
extra check, an error will be returned to prevent the code from
continuing to run.

In the meanwhile, this field will always be set to the maximal one
fetched from the capability without any change.  Removing the value
set and leave it to 0 will help to solve the potential failure.

Fixes: e1da60a8a6e9 ("common/mlx5: add DevX command to modify RQT")
Cc: matan@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>