app/testpmd: restore VXLAN-GPE support
VXLAN-GPE support was removed from testpmd recently. Drivers which are
not migrated are still using VXLAN-GPE in tests.

This commit is to restore the support for VXLAN-GPE in testpmd.

After this change, there are two command line items with the same name, ie.
"protocol". One is for the new VXLAN extensions and the other is for the
deprecated VXLAN-GPE. Add a new one that is more obvious for the new VXLAN
structure since the two have different command line context.

Fixes: da118115d95c ("app/testpmd: support matching any VXLAN field")
Signed-off-by: Gavin Li <gavinl@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>