app/crypto-perf: test queue pair priority
Updated the application to test variable queue pair priority.
A mask `--low-prio-qp-mask mask` can be set to lower the priority
of queues which are set in the mask.
This would result in lower performance for those queues.
By default the priority is set as highest.
This option is added just to verify the performance drop
of queues which have lower priority set.

Signed-off-by: Akhil Goyal <gakhil@marvell.com>
Change-Id: I88c3877f86e92a18bdb49c4421957b1c85dd10c8
Signed-off-by: 0-day Robot <robot@bytheb.org>