Revert "doc: extension of crypto event callback announced"
The previously planned update to the `rte_cryptodev_cb_fn` function
prototype is being reverted. The introduction of the
`rte_cryptodev_queue_pair_event_error_query` API provides a means for
applications to retrieve the queue pair ID that encountered an error
interrupt. This makes the addition of the `qp_id` parameter to the
`rte_cryptodev_cb_fn` function unnecessary, and as such, the proposed
extension is no longer required.

Signed-off-by: Srujana Challa <schalla@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>