bus/dpaa: add OH port mode for dpaa eth
NXP DPAA architecture supports the concept of DPAA
port as Offline Port - meaning - not connected to an actual MAC.
This is an virtual port to be used by application for exchanging data.

This property is completely driven by the device-tree. During the
DPAA bus scan, based on the platform device properties as in
device-tree, the port can be classified as OH port.

This patch add support in the driver to use dpaa eth port
in OH mode as well with DPDK applications.

Signed-off-by: Rohit Raj <rohit.raj@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>