net/ice: support optional flags in signature segment header
An optional flag field has been added to the signature segment
header. The flags field contains two flags, a valid flag, and
a last segment flag that indicates whether the segment is the
last segment that will be downloaded to firmware.

If the flag field's valid bit is NOT set, then as was done
before, assume that this is the last segment being downloaded.

However, if the flag field's valid bit IS set, then use the
last segment flag to determine if this segment is the last
segment to download.

Signed-off-by: Soumyadeep Hore <soumyadeep.hore@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>