net/ice: add support for FEC auto-detect for E830
Extends the functionality of the function responsible for
checking if the firmware supports FEC (Forward Error Correction)
disable in Auto FEC mode. It now includes an additional check to
determine if the adapter is a E830 model. With this
change, the function will enable FEC auto-detect support for
E830 adapters.

Signed-off-by: Soumyadeep Hore <soumyadeep.hore@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>