ethdev: jump to table support
Introduce new JUMP_TO_TABLE and JUMP_TO_TABLE_INDEX actions.
They allow bypassing a hierarchy of groups and going directly
to a specified flow table. That gives a user the flexibility
to jump between different priorities in a group and eliminates
the need to do a table lookup in the group hierarchy.

The JUMP_TO_TABLE action redirects a packet to a regular
flow group with the pattern-based flow rules insertion type.
The JUMP_TO_TABLE_INDEX action forwards a packet to the
specified rule index inside the index-based flow table.

The current index-based flow table doesn't do any matching
on the packet and executes the actions immediately.
Add a new index-based flow table with pattern matching.
The JUMP_TO_TABLE_INDEX can redirect a packet to another
matching criteria at the specified index in this case.

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>